refactor: move provider component and context to features/connection
