+++ client/src/features/connection/index.ts
@@ -1,0 +1,4 @@
+import ConnectionProvider from "./ConnectionProvider";
+import { ConnContext } from "./ConnContext";
+
+export { ConnContext, ConnectionProvider };
